playing around to reduce the memory consumption
This commit tries to avoid using `DBM::load_appointments` and tries to
load the appointments in a streamable fashion. The avoid using an
intermediate extended appointment hashmap.
This reduced the memory consumption in my machine from 2G to 1.4G.

This is more efficient in general but I don't get how it could be more
memory-efficient on the long run, since the intermediate hashmap is
dropped anyway at the end of `Watcher::new`.