how `iter` controlling the deallocation behaviour
To try this out, compile the binary once with `optimize` (in watcher.rs, Watcher::new)
to true and run `teosd` on a large DB and observe the memory consumption, then run
`gettowerinfo`, which is patched to do a `malloc_trim(0)` under the hood, and observe the memory consumption.

Do the same with `optimize` set to false.

With optimize set to true the tower consumes so much memory but lowers
the memory consumption when `gettowerinfo` is issued. And when optimize
is set to false, the tower consumes so much memory but doesn't lower the
memory consumption when `gettowerinfo` is issued.